+++ src/ReadoutEquipment.h
@@ -151,7 +151,7 @@ class ReadoutEquipment
   int cfgRdhCheckEnabled = 0;          // flag to enable RDH check at runtime
   int cfgRdhDumpEnabled = 0;           // flag to enable RDH dump at runtime
   int cfgRdhDumpErrorEnabled = 1;      // flag to enable RDH error log at runtime
-  int cfgRdhDumpWarningEnabled = 0;    // flag to enable RDH warning log at runtime
+  int cfgRdhDumpWarningEnabled = 1;    // flag to enable RDH warning log at runtime
   int cfgRdhUseFirstInPageEnabled = 0; // flag to enable reading of first RDH in page to populate readout headers
   //int cfgRdhCheckPacketCounterContiguous = 1; // flag to enable checking if RDH packetCounter value contiguous (done link-by-link)
   double cfgTfRateLimit = 0;           // TF rate limit, to throttle data readout
@@ -163,6 +163,11 @@ class ReadoutEquipment
 
   int processRdh(DataBlockContainerReference& nextBlock);
 
+  // data debugging to disk
+  int cfgSaveErrorPagesMax; // maximum number of pages to write to disk for debugging, in case of data error
+  std::string cfgSaveErrorPagesPath; // path to write data pages
+  int saveErrorPagesCount; // counter for number of pages dumped so far
+  
  protected:
   // get timeframe from orbit
   // orbit of TF 1 is set on first call